Deduplicate ports auto forwarded on all interfaces and localhost
